+++ packages/nesd/test/test_roms/blargg_status_test.dart
@@ -19,6 +19,7 @@ const _passing = <String>[
   'cpu_exec_space/test_cpu_exec_space_apu.nes',
   'cpu_exec_space/test_cpu_exec_space_ppuio.nes',
   'ppu_open_bus/ppu_open_bus.nes',
+  'ppu_read_buffer/test_ppu_read_buffer.nes',
   'sprdma_and_dmc_dma/sprdma_and_dmc_dma.nes',
   'sprdma_and_dmc_dma/sprdma_and_dmc_dma_512.nes',
   'cpu_reset/ram_after_reset.nes',